requirements.txt
base2048>=0.1.3
transformers>=4.40.0
datasets>=2.14.6,<2.17
colorama>=0.4.3
tqdm>=4.64.0
cohere>=4.5.1,<5
openai>=1.14.0,<2
replicate>=0.8.3
google-api-python-client>=2.0
backoff>=2.1.1
rapidfuzz>=3.0.0
jinja2>=3.1.2
nltk>=3.8.1
accelerate>=0.23.0
avidtools==0.1.2
stdlibs>=2022.10.9
langchain>=0.0.300
nemollm>=0.3.0
octoai-sdk>=0.8.0
cmd2==2.4.3
torch>=2.1.3
sentencepiece>=0.1.99
markdown>=3.4.3
numpy>=1.26.1
zalgolib>=0.2.2
ecoji>=0.1.0
deepl==1.17.0
fschat>=0.2.36
litellm>=1.33.8
jsonpath-ng>=1.6.1
# tests
pytest>=8.0
requests-mock==1.12.1
# lint
black>=22.3
pylint>=3.1.0